chore(tnode): update tnode function
